modified:   .DS_Store 	modified:   data/.DS_Store 	modified:   data/cache/0/015f9b4d23d604e917f00446d3912a10.css 	modified:   data/cache/0/015f9b4d23d604e917f00446d3912a10.css.gz 	modified:   data/cache/0/0a0ef8c7736c13ae6705bf73052a6d54.metadata 	modified:   data/cache/1/152a9d09221c6fa13e78d3ccf60e38ae.metadata 	modified:   data/cache/3/3bf118623fd96e3ca4acf1ac8c93423d.metadata 	modified:   data/cache/4/4d6b804a141dd74a0541297959106926.metadata 	modified:   data/cache/7/745138e0a6323c26f361139828e68e35.metadata 	modified:   data/cache/e/e992fb59695510dd4c5eb91d2350228e.metadata 	modified:   data/media/.DS_Store 	modified:   lib/plugins/hello/action.php 	modified:   lib/plugins/splash/action.php 	modified:   lib/plugins/video/action.php 	modified:   lib/tpl/default/design.css 	modified:   lib/tpl/default/layout.css 	modified:   lib/tpl/guu/ajaxupload.php 	deleted:    lib/tpl/guu/f4player.php 	modified:   lib/tpl/guu/jquery.php 	modified:   lib/tpl/guu/layout.css 	modified:   lib/tpl/guu/logo.html 	modified:   lib/tpl/guu/main.php 	modified:   mysql.php 	modified:   php.php

+++ mysql.php
@@ -1,4 +1,17 @@
 <?php
+
+function coded_get($str)
+{
+	$file = getFormValue($str);
+
+	$dbdir = "";
+	$ext = strrchr($file,'.');
+	$name = substr($file,0, -(strlen($ext)));
+		
+//	return $dbdir.base64_encode($name).$ext;
+	return $dbdir.$name.$ext;
+}
+
 include_once "function.php";
 
 //if ('cgi-fcgi' != php_sapi_name()) die();
@@ -7,9 +20,9 @@ if(!defined('DOKU_INC')) define('DOKU_INC',dirname(__FILE__).'/');
 require_once(DOKU_INC.'inc/init.php');
 require_once(DOKU_INC.'inc/auth.php');
 require_once(DOKU_INC.'inc/common.php');
+
 $INFO = pageinfo();
 if($INFO["isadmin"] !== TRUE) { die("You dont have access rights"); }
-
 
 $sqlsrv = "127.0.0.1";
 $sqlusr = "root";
@@ -18,9 +31,10 @@ $sqldb  = "guu_wentang";
 
 $id          = getFormValue("id"    );
 
-$thumb       = getFormValue("thumb" );
-$data        = getFormValue("data"  );
-$small_thumb = getFormValue("st"    ); // small thumb 
+$thumb       = coded_get("thumb" );
+$data        = coded_get("data"  );
+
+$small_thumb = coded_get("st"    ); // small thumb 
 
 $table		 = getFormValue("tab"   );
 
@@ -92,6 +106,10 @@ $sql = "";
 
 	if(strcmp($delete,"yes") == 0 )
 	{
+		if($table == "" || $pid == "")
+		{
+			die("DEL input data not correct");
+		}
 		$sql = "delete  from ".$table." where pid=".$pid;
 		run_sql($sql);
 		die("success");
